se agregaron validaciones y respuestas de error en todos los endpoints

+++ src/controllers/note-controller.js
@@ -5,7 +5,7 @@ const {
     get_note_service, 
     get_notes_service, 
     create_note_service, 
-    update_task_service,
+    update_note_service,
     delete_note_service
 } = require('../services/note-service');
 
@@ -16,40 +16,62 @@ const get_note_controller = async (req, res) => {
 }
 
 const get_notes_controller = async (req, res) => {
+    console.log(req.body.email);
     const user_info = await get_user_service(req.body.email);
-    const notes = await get_notes_service(user_info._id);
-    res.send(notes);
+    if (!user_info) {
+        res.send({"message": "User does not exist"});
+    } else {
+        const notes = await get_notes_service(user_info._id);
+        res.send(notes);
+    }
 }
 
 const create_notes_controller = async (req, res) => {
-    const user_info = await get_user_service(req.body.email);
-    const result = await uploadImage(req.files.image.tempFilePath);
-    let { title, text_note } = req.body
-    let data = {
-        user_id: user_info._id,
-        title,
-        text_note,
-        image: {
-            url: result.secure_url,
-            public_id: result.public_id
+    const { email } = req.body;
+    if (!email) {
+        res.status(400).json({"message": "Email is required"});
+    } else {
+        const user_info = await get_user_service(req.body.email);
+        let { title, text_note } = req.body
+        let data = {
+            user_id: user_info._id,
+            title,
+            text_note,
         }
+        if (req.files?.image) {
+            const result = await uploadImage(req.files.image.tempFilePath);
+            data.image = {
+                url: result.secure_url,
+                public_id: result.public_id
+            }
+            fs.unlinkSync(req.files.image.tempFilePath);
+        }
+        const note = await create_note_service(data);
+        res.json(note);
     }
-    const note = await create_note_service(data);
-    fs.unlinkSync(req.files.image.tempFilePath);
-    res.send(note);
 }
 
 const update_notes_controller = async (req, res) => {
-    let { id } = req.params;
-    const note = await update_task_service(id, req.body);
-    res.send(note);
+    try {
+        let { id } = req.params;
+        const note = await update_note_service(id, req.body) || {message: "Note does not exist"};
+        res.json(note);
+    } catch (error) {
+        if (error.kind === 'ObjectId') {
+            res.json({message: 'Invalid id'});
+        } else {
+            res.json({message: error.message});
+        }
+    }
 }
 
 const delete_note_controller = async (req, res) => {
     let { id } = req.params;
-    const note = await get_note_service(id);
-    await deleteImage(note.image.public_id);
-    const note_delete = await delete_note_service(id);
+    const note = await get_note_service(id) || {message: "Note does not exist"};
+    if (note.image?.public_id) {
+        await deleteImage(note.image.public_id);
+    }
+    const note_delete = await delete_note_service(id) || {message: "Note does not exist"};
     res.send(note_delete);
 }
